backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asyncpg==0.29.0
bcrypt==4.0.1
click==8.1.7
dnspython==2.6.1
ecdsa==0.19.0
email-validator==2.1.0
fastapi==0.114.1
h11==0.14.0
idna==3.8
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.9.1
pydantic_core==2.23.3
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
